Finished Section 2.3 Tokens Railroad Diagrams Added missing case in Section 4.2 Consume a token Added 'diagrams.css' file to isolate the diagrams' css Added a 'res/' folder to store the ressources Corrected an error in the number definition in Section 2.1 Definitions Fixed some typos
Note: For now, the only way I found to properly display the diagrams is to inline them into the html file.
However, this is really inconvenient, because it takes a lot of space and is confusing for new contributors.
That's why I also included the 'res/' folder and outsourced the diagrams' css to 'diagrams.css'
If anyone has any idea how to display properly the raildroad diagrams' svgs without inline them into the html file, please consider creating a Pull Request or raising an Issue on GitHub